requirements.txt
flask
gunicorn
openai
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
PyPDF2
python-dotenv
Pillow
pillow-heif
